+++ test-parts/part-g-controller_files/controller-main.js
@@ -1,34 +1,31 @@
-const cDeviceSettingsFile = require("./files/con-device_settings");
-const cDeviceClassesFile = require("./files/con-device_classes");
-const cRioFactoriesFile = require("./files/con-rio_factories");
-const cSettingsMainFile = require("./files/con-settings");
-const cNodeRedMainFile = require("./files/con-settings_red");
+const deviceSettingsTests = require("./files/con-device_settings");
+const classTests = require("./files/con-device_classes");
+const factoryTests = require("./files/con-rio_factories");
+const settingsMainTests = require("./files/con-settings");
+const redTests = require("./files/con-settings_red");
 
 
 function coordinateController()
 {
 	describe("G - Controller Files", function()
 	{
-		var cRioIndexMainFile = require("./files/con-rio_index_main");
-		var cRioIndexNodeListFile = require("./files/con-rio_index_node_list");
-		var cRioIndexNodeRegisterFile = require("./files/con-rio_index_node_reg");
-		var cRioIndexNodeRegisterInvalid = require("./files/con-rio_index_node_reg_invalid");
-		var cServiceMainFile = require("./files/con-service_main");
+		var rioIndexMainTests = require("./files/con-rio_index_main");
+		var rioIndexNodeListTests = require("./files/con-rio_index_node_list");
+		var rioIndexRegisterTests = require("./files/con-rio_index_node_reg");
+		var rioIndexRegisterInvalidTests = require("./files/con-rio_index_node_reg_invalid");
+		var serviceTests = require("./files/con-service_main");
 		
-		cDeviceSettingsFile.callTestDeviceSettings();
-		cDeviceClassesFile.callTestDeviceClasses();
-		cRioFactoriesFile.callTestRemoteIoFactories();
-		cRioIndexMainFile.callTestRemoteIoIndexMain();
-		cRioIndexNodeListFile.callTestRemoteIoIndexNodeList();
-		cRioIndexNodeRegisterFile.callTestRemoteIoIndexNodeReg();
-		cRioIndexNodeRegisterInvalid.callTestRemoteIoIndexRegisterInvalid();
-		cServiceMainFile.callTestServiceMain();
-		cSettingsMainFile.callTestControllerSettings();
-		cNodeRedMainFile.callTestNodeRedSettings();
+		deviceSettingsTests();
+		classTests();
+		factoryTests();
+		rioIndexMainTests();
+		rioIndexNodeListTests();
+		rioIndexRegisterTests();
+		rioIndexRegisterInvalidTests();
+		serviceTests();
+		settingsMainTests();
+		redTests();
 	});
 }
 
-module.exports =
-{
-	callController: coordinateController
-};
+module.exports = coordinateController;