+++ test-parts/part-g-controller_files/files/con-device_settings.js
@@ -402,7 +402,4 @@ function getSupportedLists()
 }
 
 
-module.exports =
-{
-	callTestDeviceSettings: testDeviceSettings
-};
+module.exports = testDeviceSettings;