+++ test-parts/part-g-controller_files/files/con-rio_factories.js
@@ -153,7 +153,4 @@ function callRemoteIoModuleInvalid(rioInvalidObject)
 }
 
 
-module.exports =
-{
-	callTestRemoteIoFactories: testRemoteIoFactories
-};
+module.exports = testRemoteIoFactories;