+++ test-parts/part-g-controller_files/files/con-rio_index_main.js
@@ -173,7 +173,4 @@ function getCrudFunctions()
 	return namesList;
 }
 
-module.exports =
-{
-	callTestRemoteIoIndexMain: testRemoteIoIndexMain
-};
+module.exports = testRemoteIoIndexMain;