+++ test-parts/part-g-controller_files/files/con-rio_index_node_list.js
@@ -165,7 +165,4 @@ function verifyNodeListManufacturer(mName, mError, mDevices)
 	rioSubCommonFile.callTestPropertyArrayStructure(mDevices);
 }
 
-module.exports =
-{
-	callTestRemoteIoIndexNodeList: testRemoteIoIndexNodeList
-};
+module.exports = testRemoteIoIndexNodeList;