+++ test-parts/part-g-controller_files/files/con-rio_index_node_reg.js
@@ -742,7 +742,4 @@ function coordinateGetIoPropertiesInvalidCall(invalidArg)
 }
 
 
-module.exports =
-{
-	callTestRemoteIoIndexNodeReg: testRemoteIoIndexNodeReg
-};
+module.exports = testRemoteIoIndexNodeReg;