+++ test-parts/part-g-controller_files/files/con-rio_index_node_reg_invalid.js
@@ -548,7 +548,4 @@ function handleNodeInvalidDispose()
 	});
 }
 
-module.exports =
-{
-	callTestRemoteIoIndexRegisterInvalid: testRemoteIoIndexRegisterInvalid
-};
+module.exports = testRemoteIoIndexRegisterInvalid;