+++ test-parts/part-g-controller_files/files/con-service_main.js
@@ -130,7 +130,4 @@ function handleProcessFunctions()
 	});
 }
 
-module.exports =
-{
-	callTestServiceMain: testServiceMain
-};
+module.exports = testServiceMain;