+++ test-parts/part-g-controller_files/files/con-settings.js
@@ -94,7 +94,4 @@ function handlePathProperties()
 	});
 }
 
-module.exports =
-{
-	callTestControllerSettings: testControllerSettings
-};
+module.exports = testControllerSettings;