+++ test-parts/part-g-controller_files/files/con-settings_red.js
@@ -29,7 +29,4 @@ function checkSettingsFile()
 }
 
 
-module.exports =
-{
-	callTestNodeRedSettings: testNodeRedSettings
-};
+module.exports = testNodeRedSettings;